prevent SQL injection attacks

+++ server/notes.py
@@ -23,15 +23,15 @@ def create_or_modify_note(request, is_new_note=False):
     cursor = connection.cursor()
     if not is_new_note:
         # support creating note from the UPDATE_NOTE backend, to simplify front end process
-        cursor.execute('SELECT 1 FROM NOTES WHERE folder_id="%s" and user_id="%s" and id="%s"' % (folder, user_id, id))
+        cursor.execute('SELECT 1 FROM NOTES WHERE folder_id=? and user_id=? and id=?', (folder, user_id, id))
         note_exists = cursor.fetchone()
         print('note', note_exists)
         if not note_exists or not note_exists[0]:  # looking for  either (,) or (None,)
             is_new_note = True
     if is_new_note:
-        cursor.execute(f'INSERT INTO NOTES (id, title, body, user_id, folder_id) VALUES ("{id}", "{title}", "{body}", "{user_id}", "{folder}")')
+        cursor.execute('INSERT INTO NOTES (id, title, body, user_id, folder_id) VALUES (?, ?, ?, ?, ?)', (id, title, body, user_id, folder))
     if not is_new_note:
-        cursor.execute('UPDATE NOTES SET title="%s", body="%s", folder_id="%s" where user_id="%s" and id="%s"' % (title, body, folder, user_id, id))
+        cursor.execute('UPDATE NOTES SET title=?, body=?, folder_id=? where user_id=? and id=?', (title, body, folder, user_id, id))
     connection.commit()
     connection.close()
 
@@ -55,9 +55,9 @@ def notes():
     connection.row_factory = sqlite3.Row  # results come back as dictionaries
     cursor = connection.cursor()
     if folder_id:
-        cursor.execute('SELECT * FROM NOTES WHERE user_id="%s" and folder_id="%s" ORDER BY last_modified DESC' % (user_id, folder_id,))
+        cursor.execute('SELECT * FROM NOTES WHERE user_id=? and folder_id=? ORDER BY last_modified DESC', (user_id, folder_id,))
     else:
-        cursor.execute('SELECT * FROM NOTES WHERE user_id="%s" ORDER BY last_modified DESC' % (user_id,))
+        cursor.execute('SELECT * FROM NOTES WHERE user_id=? ORDER BY last_modified DESC', (user_id,))
     results = cursor.fetchall()  # [['uadfsdf', 'title', 'body', None], []...]
     notes = []
     for result in results:
@@ -75,9 +75,10 @@ def note(id):
     if request.method == 'PUT':
         return create_or_modify_note(request)
     
+    user_id = session.get('user_id')
     connection = sqlite3.connect(DB_FILE)
     cursor = connection.cursor()
-    cursor.execute(f'SELECT * FROM NOTES WHERE id = "{id}"')
+    cursor.execute('SELECT * FROM NOTES WHERE user_id=? and id=?', (user_id, id))
     note = cursor.fetchone()
     if note is None:
         return {}  # if ID was invalid
@@ -90,14 +91,16 @@ def note(id):
 @app.route('/notes/<id>', methods=['DELETE'])
 def note_delete(id):
     tuple_ids = tuple(id.split(','))
+    user_id = session.get('user_id')
 
     connection = sqlite3.connect(DB_FILE)
     cursor = connection.cursor()
 
     if len(tuple_ids) == 1:
-        cursor.execute(f'DELETE FROM NOTES WHERE id = "{tuple_ids[0]}"')
+        cursor.execute('DELETE FROM NOTES WHERE user_id=? and id=?', (user_id, tuple_ids[0]))
     else:
-        cursor.execute(f'DELETE FROM NOTES WHERE id IN {tuple_ids}')
+        question_marks = ', '.join('?' for _ in tuple_ids)  # aka '?, ?, ?' if 3 id's passed
+        cursor.execute(f'DELETE FROM NOTES WHERE user_id=? and id IN ({question_marks})', (user_id, *tuple_ids))
 
     connection.commit()
     connection.close()
